+++ CHANGES.md
@@ -9,6 +9,7 @@ Changes in 5.5
 - SQL REPORT_CARD_GRADES table: Change title column type to character varying(5) in rosariosis.sql, Update.fnc.php & ReportCardGrades.php
 - Fix required check for TextInput(), PasswordInput() & SelectInput() in Inputs.php
 - Make Final Grading Percentages inputs required in Configuration.php
+- Input type number for Breakoff & Final Grade Percentages in Configuration.php
 
 Changes in 5.4.5
 ----------------